Switch from fork/exec to posix_spawn
- Use posix_spawn() for cleaner child process creation
- posix_spawn_file_actions_addchdir_np for working directory
- Fix float-to-int warnings in ComposeComponent

+++ juce_cmp/juce_cmp/ComposeComponent.cpp
@@ -70,7 +70,7 @@ void ComposeComponent::paint(juce::Graphics& g)
         }
 
         g.drawImage(loadingPreview_,
-                    drawX, drawY, drawWidth, drawHeight,
+                    (int)drawX, (int)drawY, (int)drawWidth, (int)drawHeight,
                     0, 0, loadingPreview_.getWidth(), loadingPreview_.getHeight());
     }
 }
